Create functioning
As it turns out, changes I made to the schema on my back end have not
stuck. The Results array is still going by questions and the prompt string
is still going by title. This needs to be rectified, but it is heartening
to see that the DB can be successfully touched from the front end.