Delete functionality built, but fails
Which is utterly mystifying, as it is mechanically the same as GET, taking
in an ID and deleting it rather than retrieving it. Probably the strangest
part about this is that the error being thrown is a 404, which is patently
false, as you can check it in the server.

Entirely possible that back end delete simply functions differently with
mongo. Going to look into it.