Found source of error, but mystified as to how to fix it
The problem looks like the fact that AJAX is improperly constructing
the URL, sending the wrong thing to the server.

Frankly speaking, I'm mystified as to how I'd solve this. I've been
fiddling around on both the back end and the front end and am no closer
to a solution.